64int move -> 32int move + basic search

+++ src/main/java/fr/flwrian/aspira/board/BoardHistoryStack.java
@@ -13,7 +13,7 @@ public class BoardHistoryStack {
         }
     }
 
-    public void push(Board board, long move) {
+    public void push(Board board, int move) {
         if (top + 1 >= stack.length) {
             throw new RuntimeException("BoardHistoryStack overflow");
         }
